Adding router rewrite files (keeping old for reference)
