+++ routing/route_collection.go
@@ -1,0 +1,8 @@
+package routing
+
+type RouteCollection struct {
+}
+
+func NewRouteCollection() *RouteCollection {
+	return new(RouteCollection)
+}
