+++ routing/route_collection_test.go
@@ -1,0 +1,13 @@
+package routing
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestNewRouteCollection(t *testing.T) {
+	rc := NewRouteCollection()
+
+	assert.IsType(t, new(RouteCollection), rc)
+}
